requirements.txt
streamlit
sympy
matplotlib
numpy
pillow
pytesseract
python-docx